requirements.txt
# Web Framework
Flask>=3.0.0
Flask-Cors>=4.0.0

# OCR and Document Processing
python-doctr[torch]>=0.10.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
pdf2image>=1.16.3
pypdfium2>=4.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Machine Learning & Deep Learning
torch>=2.2.2+cpu
torchvision>=0.17.2+cpu

# Text Processing & Pattern Matching
regex>=2023.0.0
python-Levenshtein>=0.21.0
rapidfuzz>=3.0.0

# Database
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
PyMySQL>=1.1.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
fonttools>=4.40.0
python-dotenv>=1.0.0
requests>=2.31.0

